pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.159"
invoke = "^2.2.0"
pandas = "^2.2.2"
pytz = "^2024.1"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
tqdm = "^4.66.5"
pypdf2 = "^3.0.1"
moto = "^5.0.12"
natsort = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
